+++ checksum-Elixir.FinanceRustler.Solver.exs
@@ -1,0 +1,4 @@
+%{
+  "libfinance_rustler-v0.1.0-nif-2.15-aarch64-unknown-linux-gnu.so.tar.gz" => "sha256:def2e62e09f497f74fb92d2e58a351b19853cc76a2f15d3f2cd4d39b6405fc24",
+  "libfinance_rustler-v0.1.0-nif-2.15-x86_64-unknown-linux-gnu.so.tar.gz" => "sha256:379c38f0737bae130217c621b0bec683b3ad0bbd1d0d533b32c292c20e5242c1",
+}